fixed ref so it applies per line
